requirements.txt
streamlit
transformers
torch
pymupdf
reportlab
openai-whisper
sentencepiece